+++ apps/formio-server/src/hooks/settings.js
@@ -65,6 +65,9 @@ module.exports = function(app) {
         if (req.method === 'POST' && alias.match(/\/validate$/)) {
           return alias.replace(/\/validate$/, '');
         }
+        else if (req.url.includes('/pdf-proxy')) {
+          return false;
+        }
         else {
           return alias;
         }
@@ -261,8 +264,11 @@ module.exports = function(app) {
         if (req.url.indexOf(`/project/${req.projectId}/manage`) === 0) {
           return true;
         }
+        if (req.url.includes('/pdf-proxy')) {
+          return true;
+        }
 
-       if (req.headers['x-jwt-token'] && req.user) {
+        if (req.headers['x-jwt-token'] && req.user) {
           const whitelist2fa = ['/2fa/generate', '/2fa/represent', '/2fa/turn-on', '/2fa/turn-off'];
           const url = req.url.split('?')[0];
           const is2fa = _.some(whitelist2fa, (path) => {
@@ -276,7 +282,7 @@ module.exports = function(app) {
           if (is2fa) {
             return true;
           }
-       }
+        }
 
         if (req.method !== 'GET') {
           return false;