+++ apps/formio-server/Changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased: 8.2.0-rc.3]
 ### Fixed
  - FIO-7060: Fixing the pdf-proxy to download and authenticate correctly
+ - FIO-7103: fixed an issue where the webhook action would strip query parameters
 
 ## 8.2.0-rc.2
 ### Changed