+++ apps/formio-server/Changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
  - FIO-7060: Fixing the pdf-proxy to download and authenticate correctly
  - FIO-7103: fixed an issue where the webhook action would strip query parameters
+ - FIO-7136, FIO-7137: Fixed issues with pdf-proxy authentication.
 
 ## 8.2.0-rc.2
 ### Changed