+++ apps/formio-server/Changelog.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased: 8.2.0-rc.3]
+## 8.2.0-rc.3
+### Changed
+ - Updated formio-app@8.2.0-rc.3
+ - Updated formiojs@4.16.0-rc.7
+ - Updated @formio/premium@1.20.0-rc.2
+ - Updated formio@3.2.0-rc.3
+ - Updated formio-workers@1.18.0-rc.3
+
 ### Fixed
  - FIO-7060: Fixing the pdf-proxy to download and authenticate correctly
  - FIO-7103: fixed an issue where the webhook action would strip query parameters